requirements.txt
aiofiles==23.1.0
aiosmtplib==2.0.1
aiosqlite==0.19.0
alembic==1.9.4
amqp==5.1.1
anyio==3.6.2
APScheduler==3.10.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.2.0
billiard==3.6.4.0
bitly-api-python==1.0.3
blinker==1.6.2
celery==5.2.7
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coverage==7.2.2
cryptography==40.0.1
Deprecated==1.2.13
dnspython==2.3.0
ecdsa==0.18.0
email-validate==1.1.2
email-validator==1.3.1
exceptiongroup==1.1.0
fastapi==0.94.0
fastapi-mail==1.2.7
fastapi-pagination==0.11.4
fastapi_amis_admin==0.5.5
filelock==3.9.0
flower==1.2.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
humanize==4.6.0
idna==2.10
iniconfig==2.0.0
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
nanoid==2.0.0
ngrok==1.0.0
numpy==1.26.2
packaging==23.0
pandas==2.1.3
password-strength==0.0.3.post2
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
pyactiveresource==1.0.1
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
pymailgun==1.1
pyngrok==7.1.2
pytest==7.2.1
pytest-cov==4.0.0
pytest-env==0.8.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
redbird==0.7.1
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.40.0
sqlalchemy2-stubs==0.0.2a34
sqlalchemy_database==0.1.0
sqlmodel==0.0.8
starlette==0.26.0.post1
text-unidecode==1.3
tomli==2.0.1
tornado==6.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.14
uvicorn==0.21.0
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0